+++ packages/sdk/src/internal/sdk/sdk-factory.ts
@@ -1,3 +1,43 @@
+import { Auth } from '@auth/auth.interface';
+import { DialectWalletAdapterEd25519TokenSigner } from '@auth/signers/ed25519-token-signer';
+import { DialectWalletAdapterSolanaTxTokenSigner } from '@auth/signers/solana-tx-token-signer';
+import type { TokenProvider } from '@auth/token-provider';
+import { DEFAULT_TOKEN_LIFETIME } from '@auth/token-provider';
+import { TokenStore } from '@auth/token-store';
+import type { DappAddresses, DappMessages, Dapps } from '@dapp/dapp.interface';
+import { DappsImpl } from '@dapp/internal/dapp';
+import { DappAddressesFacade } from '@dapp/internal/dapp-addresses-facade';
+import { DappMessagesFacade } from '@dapp/internal/dapp-messages-facade';
+import { DataServiceDappNotificationSubscriptions } from '@dapp/internal/dapp-notification-subscriptions';
+import { DataServiceDappNotificationTypes } from '@dapp/internal/dapp-notification-types';
+import { DataServiceDappAddresses } from '@dapp/internal/data-service-dapp-addresses';
+import { DataServiceDappMessages } from '@dapp/internal/data-service-dapp-messages';
+import { SolanaDappAddresses } from '@dapp/internal/solana-dapp-addresses';
+import { SolanaDappMessages } from '@dapp/internal/solana-dapp-messages';
+import { DataServiceApi } from '@data-service-api/data-service-api';
+import type { DataServiceDappNotificationSubscriptionsApi } from '@data-service-api/data-service-dapp-notification-subscriptions-api';
+import type { DataServiceDappNotificationTypesApi } from '@data-service-api/data-service-dapp-notification-types-api';
+import type { DataServiceDappsApi } from '@data-service-api/data-service-dapps-api';
+import type { DataServiceDialectsApi } from '@data-service-api/data-service-dialects-api';
+import { programs } from '@dialectlabs/web3';
+import { EncryptionKeysStore } from '@encryption/encryption-keys-store';
+import { EncryptionKeysProvider } from '@encryption/internal/encryption-keys-provider';
+import type { IdentityResolver } from '@identity/identity.interface';
+import {
+  AggregateSequentialIdentityResolver,
+  FirstFoundFastIdentityResolver,
+  FirstFoundIdentityResolver,
+} from '@identity/internal/identity-resolvers';
+import { DataServiceMessaging } from '@messaging/internal/data-service-messaging';
+import {
+  MessagingBackend,
+  MessagingFacade,
+} from '@messaging/internal/messaging-facade';
+import { createDialectProgram } from '@messaging/internal/solana-dialect-program-factory';
+import { SolanaMessaging } from '@messaging/internal/solana-messaging';
+import type { Messaging } from '@messaging/messaging.interface';
+import type { Program } from '@project-serum/anchor';
+import { IllegalArgumentError } from '@sdk/errors';
 import {
   Backend,
   Config,
@@ -5,47 +45,14 @@ import {
   DialectCloudConfig,
   DialectSdk,
   DialectSdkInfo,
+  IdentityConfig,
   SolanaConfig,
 } from '@sdk/sdk.interface';
-import { programs } from '@dialectlabs/web3';
-import { DialectWalletAdapterWrapper } from '@wallet-adapter/dialect-wallet-adapter-wrapper';
-import { DataServiceApi } from '@data-service-api/data-service-api';
-import { DataServiceMessaging } from '@messaging/internal/data-service-messaging';
-import {
-  MessagingBackend,
-  MessagingFacade,
-} from '@messaging/internal/messaging-facade';
 import { PublicKey } from '@solana/web3.js';
-import { createDialectProgram } from '@messaging/internal/solana-dialect-program-factory';
-import { Duration } from 'luxon';
-import { SolanaMessaging } from '@messaging/internal/solana-messaging';
-import { DialectWalletAdapterEd25519TokenSigner } from '@auth/signers/ed25519-token-signer';
-import { IllegalArgumentError } from '@sdk/errors';
-import type { DappAddresses, DappMessages, Dapps } from '@dapp/dapp.interface';
-import type { Program } from '@project-serum/anchor';
-import { DappsImpl } from '@dapp/internal/dapp';
-import { DappAddressesFacade } from '@dapp/internal/dapp-addresses-facade';
-import { SolanaDappAddresses } from '@dapp/internal/solana-dapp-addresses';
-import { DataServiceDappAddresses } from '@dapp/internal/data-service-dapp-addresses';
-import type { DataServiceDialectsApi } from '@data-service-api/data-service-dialects-api';
-import type { DataServiceDappsApi } from '@data-service-api/data-service-dapps-api';
+import { DialectWalletAdapterWrapper } from '@wallet-adapter/dialect-wallet-adapter-wrapper';
 import { DataServiceWallets } from '@wallet/internal/data-service-wallets';
-import type { Messaging } from '@messaging/messaging.interface';
 import type { Wallets } from '@wallet/wallet.interface';
-import { EncryptionKeysProvider } from '@encryption/internal/encryption-keys-provider';
-import { EncryptionKeysStore } from '@encryption/encryption-keys-store';
-import { TokenStore } from '@auth/token-store';
-import { DappMessagesFacade } from '@dapp/internal/dapp-messages-facade';
-import { SolanaDappMessages } from '@dapp/internal/solana-dapp-messages';
-import { DataServiceDappMessages } from '@dapp/internal/data-service-dapp-messages';
-import { DialectWalletAdapterSolanaTxTokenSigner } from '@auth/signers/solana-tx-token-signer';
-import { DataServiceDappNotificationTypes } from '@dapp/internal/dapp-notification-types';
-import type { DataServiceDappNotificationTypesApi } from '@data-service-api/data-service-dapp-notification-types-api';
-import { DataServiceDappNotificationSubscriptions } from '@dapp/internal/dapp-notification-subscriptions';
-import type { DataServiceDappNotificationSubscriptionsApi } from '@data-service-api/data-service-dapp-notification-subscriptions-api';
-import { Auth } from '@auth/auth.interface';
-import type { TokenProvider } from '@auth/token-provider';
-import { DEFAULT_TOKEN_LIFETIME } from '@auth/token-provider';
+import { Duration } from 'luxon';
 
 interface InternalConfig extends Config {
   wallet: DialectWalletAdapterWrapper;
@@ -57,6 +64,7 @@ export class InternalDialectSdk implements DialectSdk {
     readonly threads: Messaging,
     readonly dapps: Dapps,
     readonly wallet: Wallets,
+    readonly identity: IdentityResolver,
   ) {}
 }
 
@@ -112,6 +120,8 @@ export class DialectSdkFactory {
       dataServiceApi.walletNotificationSubscriptions,
       dataServiceApi.pushNotificationSubscriptions,
     );
+    const identity = this.createIdentityResolver(config.identity);
+
     return new InternalDialectSdk(
       {
         apiAvailability: config.wallet,
@@ -125,6 +135,7 @@ export class DialectSdkFactory {
       messaging,
       dapps,
       wallet,
+      identity,
     );
   }
 
@@ -240,11 +251,28 @@ Solana settings:
     );
   }
 
+  private createIdentityResolver(config: IdentityConfig): IdentityResolver {
+    if (config.strategy === 'first-found') {
+      return new FirstFoundIdentityResolver(config.resolvers);
+    }
+    if (config.strategy === 'first-found-fast') {
+      return new FirstFoundFastIdentityResolver(config.resolvers);
+    }
+    if (config.strategy === 'aggregate-sequential') {
+      return new AggregateSequentialIdentityResolver(config.resolvers);
+    }
+
+    throw new IllegalArgumentError(
+      `Unknown identity strategy ${config.strategy}`,
+    );
+  }
+
   private initializeConfig(): InternalConfig {
     const environment = this.config.environment ?? 'production';
     const wallet = DialectWalletAdapterWrapper.create(this.config.wallet);
     const backends = this.initializeBackends();
     const encryptionKeysStore = this.createEncryptionKeysStore();
+    const identity = this.createIdentityConfig();
     return {
       environment,
       wallet,
@@ -252,6 +280,7 @@ Solana settings:
       solana: this.initializeSolanaConfig(),
       encryptionKeysStore,
       backends,
+      identity,
     };
   }
 
@@ -414,5 +443,26 @@ Solana settings:
       internalConfig.rpcUrl = this.config.solana.rpcUrl;
     }
     return internalConfig;
+  }
+
+  private createIdentityConfig(): IdentityConfig {
+    const identityConfig: IdentityConfig = {
+      strategy: 'first-found',
+      resolvers: [],
+    };
+
+    if (!this.config.identity) {
+      return identityConfig;
+    }
+
+    if (this.config.identity.strategy) {
+      identityConfig.strategy = this.config.identity.strategy;
+    }
+
+    if (this.config.identity.resolvers) {
+      identityConfig.resolvers = this.config.identity.resolvers;
+    }
+
+    return identityConfig;
   }
 }