+++ packages/sdk/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - feature: configurable token lifetime
 - feature: summary across all threads (currently: amount of unread messages)
+- feature: add identity providers
 
 ## [0.6.5] - 2022-08-15